fix(vite): keep node builtins out of the client locale module
The virtual:sveltepress/locale module is imported by client code, but
src/locale.ts imported node:fs/node:path at module top level for build-time
route scanning. Vite externalizes node builtins for the browser, so hydration
crashed with "Module node:fs has been externalized" and SvelteKit rendered
the 500 Internal Error page on every docs-site page.

Split the module: filesystem scanning (resolveLocalesConfig, scanLocaleRoutes,
...) moved to src/locale-scan.ts; locale.ts now contains only pure string
helpers with zero node imports. Add a regression test asserting the client
module stays node-free and the scan module keeps the fs logic.
