+++ docs/superpowers/specs/2026-09-01-i18n-support.md
@@ -113,6 +113,16 @@ Proof commands run on the unmodified tree; all green unless noted:
 9. `grep -r "sveltepress-docs-site-zh\|sveltepress-docs-site-bn" .github/workflows/deploy-netlify.yml` → matches (expected; removed by T7)
 10. `pnpm test` → all suites pass (scripts, cli, vite, theme-default, theme-blog), exit 0
 
+## Post-ship Fix (2026-09-01)
+
+**Bug:** Visiting any docs-site page in a real browser showed the SvelteKit 500 "Internal Error" page. SSR returned 200 (curl tests passed) but client-side hydration crashed with `Error: Module "node:fs" has been externalized for browser compatibility. Cannot access "node:fs.existsSync" in client code.`
+
+**Root cause:** `packages/vite/src/locale.ts` (exported as `@sveltepress/vite/locale`) imported `node:fs`/`node:path` at module top level for build-time route scanning. The `virtual:sveltepress/locale` virtual module imports its helpers from `@sveltepress/vite/locale` in **client code**, so the whole module — including the `node:fs` import — was bundled into the browser. Node built-ins are externalized by Vite for browser compatibility, so executing `existsSync` threw at hydration.
+
+**Fix:** Split the module. Build-time filesystem scanning (`resolveLocalesConfig`, `scanLocaleRoutes`, `localeRoutesDir`, `collectPageFiles`, `deriveRoute`, `normalizeRoute`) moved to new `packages/vite/src/locale-scan.ts`; `locale.ts` now contains only pure string functions (`resolveLocale`, `resolveLocalizedPath`, `resolveLocaleSwitch`) with zero Node.js imports. `plugin.ts` imports `resolveLocalesConfig` from `./locale-scan.js`; `build.config.ts` gains a `src/locale-scan` entry. Added regression test `packages/vite/__tests__/locale-client-safety.test.ts` asserting `locale.ts` has no `node:` imports and `locale-scan.ts` keeps the filesystem scanning.
+
+**Verification:** full vite suite 24 files / 143 tests passed; theme-default 17 files / 86 tests passed; Playwright (real Chromium) renders `/`, `/zh/`, `/bn/`, guide/reference/whats-new pages with no page errors, desktop and mobile language switchers navigate `/guide/introduction/` → `/zh/guide/introduction/` preserving the page.
+
 ## Plan
 
 - [x] **T1 — Core `locales` option, locale resolution, and locale virtual module.** Delivers: `sveltepress({ locales })` accepted with per-locale theme options; prefix-based locale matching; `virtual:sveltepress/locale` exposing the resolved current locale and path helpers; no behavior change without the option. Blocked by: nothing.
